requirements.txt
fastapi
uvicorn[standard]
beanie
motor
pydantic
pydantic[email]
email-validator
python-dotenv
numpy